ADD: added trending, new and hot section to search page

+++ app/src/main/java/com/hapramp/activity/DetailedActivity.java
@@ -171,7 +171,7 @@ public class DetailedActivity extends AppCompatActivity {
     private void collectExtras() {
 
         post = getIntent().getExtras().getParcelable(Constants.EXTRAA_KEY_POST_DATA);
-        currentCommentUrl = String.format(getResources().getString(R.string.commentUrl), Integer.valueOf(post.id));
+        currentCommentUrl = String.format(getResources().getString(R.string.commentUrl), Long.valueOf(post.id));
         progressDialog = new ProgressDialog(this);
 
     }
